// src/Constants/ResponseStatus.d.ts
declare enum ResponseStatus {
    BUCKAROO_STATUSCODE_SUCCESS = "190",
    BUCKAROO_STATUSCODE_FAILED = "490",
    BUCKAROO_STATUSCODE_VALIDATION_FAILURE = "491",
    BUCKAROO_STATUSCODE_TECHNICAL_ERROR = "492",
    BUCKAROO_STATUSCODE_REJECTED = "690",
    BUCKAROO_STATUSCODE_WAITING_ON_USER_INPUT = "790",
    BUCKAROO_STATUSCODE_PENDING_PROCESSING = "791",
    BUCKAROO_STATUSCODE_WAITING_ON_CONSUMER = "792",
    BUCKAROO_STATUSCODE_PAYMENT_ON_HOLD = "793",
    BUCKAROO_STATUSCODE_CANCELLED_BY_USER = "890",
    BUCKAROO_STATUSCODE_CANCELLED_BY_MERCHANT = "891",
    BUCKAROO_AUTHORIZE_TYPE_CANCEL = "I014",
    BUCKAROO_AUTHORIZE_TYPE_ACCEPT = "I013",
    BUCKAROO_AUTHORIZE_TYPE_GROUP_TRANSACTION = "I150"
}
export default ResponseStatus;